requirements.txt
gunicorn>=21.2,<22.0
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
drf-spectacular>=0.28.1,<1.0
celery>=5.3,<6.0
redis>=5.0,<6.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
pytest>=7.0,<8.0
pytest-cov>=7.0.0,<8.0
pytest-django>=4.5,<5.0
factory-boy>=3.2,<4.0
ruff>=0.13.2,<0.14.0
